Add ACL

+++ functions/rest/entry.js
@@ -1,20 +1,34 @@
-const {JsonApiRequest, JsonApiResponse} = require('../../lib/specs/jsonApi')
-const {EntryNotFoundError, ModelNotFoundError} = require('../../lib/errors')
-const getEventToken = require('../../lib/acl/getEventToken')
+const {
+  EntryNotFoundError,
+  ForbiddenError,
+  ModelNotFoundError,
+  UnauthorizedError
+} = require('../../lib/errors')
+const FieldSet = require('../../lib/fieldSet')
+const JsonApiRequest = require('../../lib/specs/jsonApi/request')
+const JsonApiResponse = require('../../lib/specs/jsonApi/response')
 const modelFactory = require('../../lib/modelFactory')
 const schemaStore = require('../../lib/schemaStore')
-const validateAccess = require('../../lib/acl/validateAccess')
 
 module.exports.delete = async (req, res) => {
   try {
     const modelName = req.url.getPathParameter('modelName')
-    const schema = schemaStore.get(modelName)
+    const schema = schemaStore.get(modelName, true)
 
     if (!schema) {
       throw new ModelNotFoundError({name: modelName})
     }
 
-    const Model = modelFactory(modelName, schema)
+    const Model = modelFactory(schema.name, schema)
+    const access = await Model.getAccessForUser({
+      accessType: 'delete',
+      user: req.user
+    })
+
+    if (access.isDenied()) {
+      throw req.user ? new ForbiddenError() : new UnauthorizedError()
+    }
+
     const id = req.url.getPathParameter('id')
     const {deleteCount} = await Model.delete({id})
 
@@ -22,20 +36,18 @@ module.exports.delete = async (req, res) => {
       throw new EntryNotFoundError({id})
     }
 
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       statusCode: 200,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   } catch (errors) {
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       errors,
       request: this,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   }
@@ -44,21 +56,29 @@ module.exports.delete = async (req, res) => {
 module.exports.get = async (req, res) => {
   try {
     const modelName = req.url.getPathParameter('modelName')
-    const schema = schemaStore.get(modelName)
+    const schema = schemaStore.get(modelName, true)
 
     if (!schema) {
       throw new ModelNotFoundError({name: modelName})
     }
 
-    // const access = await validateAccess({
-    //   ...getEventToken(req),
-    //   accessType: 'read',
-    //   resource: `model:${modelName}`
-    // })
+    const Model = modelFactory(schema.name, schema)
+    const access = await Model.getAccessForUser({
+      accessType: 'read',
+      user: req.user
+    })
 
-    const Model = modelFactory(modelName, schema)
+    if (access.isDenied()) {
+      throw req.user ? new ForbiddenError() : new UnauthorizedError()
+    }
+
     const id = req.url.getPathParameter('id')
+    const urlFieldSet = (req.url.getQueryParameter('fields', {
+      isCSV: true
+    }) || {})[schema.name]
     const entry = await Model.findOneById({
+      fieldSet: FieldSet.intersect(access.fields, urlFieldSet),
+      filter: access.filter,
       id
     })
 
@@ -67,32 +87,28 @@ module.exports.get = async (req, res) => {
     }
 
     const request = new JsonApiRequest(req)
-    const referencesHash = {}
 
     await request.resolveReferences({
       entries: [entry],
       includeMap: req.url.getQueryParameter('include', {
         isCSV: true,
         isDotPath: true
-      }),
-      referencesHash
+      })
     })
 
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       entries: entry,
-      includedReferences: Object.values(referencesHash),
+      includedReferences: Object.values(request.references),
       includeTopLevelLinks: true,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   } catch (errors) {
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       errors,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   }
@@ -101,40 +117,45 @@ module.exports.get = async (req, res) => {
 module.exports.patch = async (req, res) => {
   try {
     const modelName = req.url.getPathParameter('modelName')
-    const schema = schemaStore.get(modelName)
+    const schema = schemaStore.get(modelName, true)
 
     if (!schema) {
       throw new ModelNotFoundError({name: modelName})
     }
 
-    const Model = modelFactory(modelName, schema)
+    const Model = modelFactory(schema.name, schema)
+    const access = await Model.getAccessForUser({
+      accessType: 'update',
+      user: req.user
+    })
+
+    if (access.isDenied()) {
+      throw req.user ? new ForbiddenError() : new UnauthorizedError()
+    }
+
     const id = req.url.getPathParameter('id')
     const request = new JsonApiRequest(req)
     const update = await request.getEntryFieldsFromBody()
     const entry = await Model.update({id, update})
-    const referencesHash = {}
 
     await request.resolveReferences({
       entries: [entry],
-      includeMap: req.url.getQueryParameter('include'),
-      referencesHash
+      includeMap: req.url.getQueryParameter('include')
     })
 
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       entries: entry,
-      includedReferences: Object.values(referencesHash),
+      includedReferences: Object.values(request.references),
       includeTopLevelLinks: true,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   } catch (errors) {
-    const response = new JsonApiResponse({
+    const {body, statusCode} = await JsonApiResponse.toObject({
       errors,
       url: req.url
     })
-    const {body, statusCode} = await response.toObject()
 
     res.status(statusCode).json(body)
   }
